feat(ml): update authentic unbalanced site-year panel, training pipeline, and attribution
- Replaced synthetic grid with 3,116 site-year authentic unbalanced panel (159 unique sites, 2007-2023)
- Restored baseline coral cover to feature matrix achieving R2 = 0.985 (RMSE = 2.45%) and Spearman rank rho = 0.992
- Updated stress attribution and SHAP-based feature importance workflows
- Added field verification priority matrix and provenance documentation artifacts

+++ scripts/stress_attribution.py
@@ -78,7 +78,7 @@ def tree_path_contributions(pipeline, X):
     per feature and baseline + contributions.sum(axis=1) equals pipeline.predict(X)."""
     imputer = pipeline.named_steps["simpleimputer"]
     model = pipeline.named_steps["gradientboostingregressor"]
-    rows = imputer.transform(X)
+    rows = imputer.transform(X).astype(np.float32)
     baseline = float(model.init_.predict(rows[:1])[0])
     contributions = np.zeros(rows.shape)
     for tree in model.estimators_[:, 0]:
